eal: add ticket based reader writer lock
This patch implements a reader/writer ticket lock because the
current DPDK reader/writer lock will starve writers when
presented with a stream of readers.

This lock type acts like rte_rwlock() but uses a ticket algorithm
and is therefore fair for multiple writers and readers. It acts
like the existing DPDK ticket and MCS lock but supports
reader/writer semantics.

It is referred to as "Simple, non-scalable, fair reader-writer lock"
in the MCS paper from PPoP '91.

The tests are just a clone of existing rte_rwlock with test
and function names changed. So the new ticket rwlocks should be drop
in replacement for most users.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>