examples/l3fwd: add FIB infrastructure
The purpose of this commit is to add the necessary function calls
and supporting infrastructure to allow the Forwarding Information Base
(FIB) library to be integrated into the l3fwd sample app.
The flag '-F' has been added to the applications options to allow
the user to specify that l3fwd uses FIB as its lookup method.

Signed-off-by: Conor Walsh <conor.walsh@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>