net/mlx5: fix UAR allocation diagnostics messages
Depending on kernel capabilities and rdma-core version the mapping
of UAR (User Access Region) of desired memory caching type (non-cached
or write combining). The PMD implements the flexible strategy of
UAR mapping, alternating the type of caching to succeed. During
this process the failure diagnostics messages are emitted. These
message are merely diagnostics ones and the logging level should be
lowered to DEBUG.

Fixes: a0bfe9d56f74 ("net/mlx5: fix UAR memory mapping type")
Cc: stable@dpdk.org

Signed-off-by: Viacheslav Ovsiienko <viacheslavo@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>