rte_metrics: unconditionally exports rte_metrics_tel_xxx functions
This patch allows the same set of rte_metrics_tel_* functions to be
exported no matter JANSSON is available or not, by doing following:
1.	Leverage dpdk_conf to set configuration flag RTE_HAVE_JANSSON
when Jansson dependency is found.
2.	In rte_metrics_telemetry.c, leverage RTE_HAVE_JANSSON to handle the
case when JANSSON is not available by adding stubs for all the instances.
3.	In meson.build, per dpdk\doc\guides\rel_notes\release_20_05.rst,
it is claimed that "Telemetry library is no longer dependent on the
external Jansson library, which allows Telemetry be enabled by default.",
thus make the deps and includes of Telemetry as not conditional anymore.

V2 changes:
    Address comments from Dmitry Kozlyuk <dmitry.kozliuk@gmail.com> and
    Bruce Richardson <bruce.richardson@intel.com>:
        - Set dpdk.conf RTE_HAS_JANSSON in metrics meson.build
        - Reduce #ifdef RTE_HAS_JANSSON blocks

V3 changes:
    Address comment from Bruce Richardson <bruce.richardson@intel.com>:
        - Remove redundant includes line on librte_telemetry

Signed-off-by: Jie Zhou <jizh@microsoft.com>
Signed-off-by: Jie Zhou <jizh@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>