doc: update minimum supported kernel
The DPDK project is only committed to supporting upstream kernel
versions that are still in support.
The 3.16 kernel has reached End Of Life (in June 2020).
The next LTS kernel is 4.19 and is supported until December 2024.

This does not change the existing policy that distribution kernels
that are older are still supported.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>